requirements.txt
streamlit
flax
jax
jaxlib
optax
tensorflow
pillow
numpy